+++ CHANGELOG.md
@@ -25,6 +25,6 @@ All notable changes to this project will be documented in this file.
 
 - Initial release
 
-[unreleased]: https://github.com/agroqirax/timberborn-controller-support/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/agroqirax/timberborn-controller-support/compare/v1.1.2.0.2...HEAD
 [1.1.2.0.2]: https://github.com/agroqirax/timberborn-controller-support/releases/tag/v1.1.2.0.2
 [1.1.2.0.1]: https://github.com/agroqirax/timberborn-controller-support/releases/tag/1.1.2.0.1